mqtt发送报警

+++ sipa-boot-common/sipa-boot-common-iot/src/main/java/com/sipa/boot/iot/protocol/mqtt/MqttHeader.java
@@ -16,7 +16,7 @@ public class MqttHeader {
 
     private String timeStamp;
 
-    private Integer index;
+    private Long index;
 
     private String function;
